create MetersToFeetConverter
A simple java program written to ask the user to decide if they want to convert meters to feet, or feet to meters as well as how many of the unit they wish to convert. The input is stored in a variable and used it to make the calculation, which is then printed to the user.